refactor: move length-constrained tuple family into its own file
Split the {Fixed,Min,Max,Bounded}LengthTuple family (and their Mutable*
variants) out of src/tuple-and-list/array.mts into
src/tuple-and-list/length-constrained-tuple.mts, mirroring
length-constrained-string.mts and length-constrained-array.mts.
array.mts now only contains the general array utilities
(NonEmptyArray, MutableNonEmptyArray, ArrayElement).

No public API change: all exports still go through the generated index
files. The README type listing is regenerated accordingly.
