+++ README.md
@@ -483,6 +483,11 @@ For detailed information on all types, see the [Full API Reference](./docs/READM
     - [ChangeBaseBrand](./src/branded-types/brand.mts#L220)
     - [IntersectBrand](./src/branded-types/brand.mts#L247)
     - [NormalizeBrandUnion](./src/branded-types/brand.mts#L272)
+- src/branded-types/predefined-arrays/length-constrained-array.mts
+    - [MaxLengthArray](./src/branded-types/predefined-arrays/length-constrained-array.mts#L64)
+    - [MinLengthArray](./src/branded-types/predefined-arrays/length-constrained-array.mts#L115)
+    - [BoundedLengthArray](./src/branded-types/predefined-arrays/length-constrained-array.mts#L153)
+    - [FixedLengthArray](./src/branded-types/predefined-arrays/length-constrained-array.mts#L188)
 - src/branded-types/predefined-numbers/bigint.mts
     - [BigInt64](./src/branded-types/predefined-numbers/bigint.mts#L18)
     - [BigUint64](./src/branded-types/predefined-numbers/bigint.mts#L33)
@@ -702,20 +707,21 @@ For detailed information on all types, see the [Full API Reference](./docs/READM
     - [RecordValueAtPath](./src/record/record-path.mts#L135)
     - [RecordValueAtPathWithIndex](./src/record/record-path.mts#L159)
 - src/tuple-and-list/array.mts
-    - [MutableNonEmptyArray](./src/tuple-and-list/array.mts#L15)
-    - [NonEmptyArray](./src/tuple-and-list/array.mts#L27)
-    - [ArrayElement](./src/tuple-and-list/array.mts#L40)
-    - [FixedLengthTuple](./src/tuple-and-list/array.mts#L54)
-    - [MutableFixedLengthTuple](./src/tuple-and-list/array.mts#L64)
-    - [MutableMinLengthTuple](./src/tuple-and-list/array.mts#L83)
-    - [MinLengthTuple](./src/tuple-and-list/array.mts#L98)
-    - [BoundedLengthTuple](./src/tuple-and-list/array.mts#L145)
-    - [MutableBoundedLengthTuple](./src/tuple-and-list/array.mts#L161)
-    - [MaxLengthTuple](./src/tuple-and-list/array.mts#L185)
-    - [MutableMaxLengthTuple](./src/tuple-and-list/array.mts#L196)
+    - [MutableNonEmptyArray](./src/tuple-and-list/array.mts#L12)
+    - [NonEmptyArray](./src/tuple-and-list/array.mts#L24)
+    - [ArrayElement](./src/tuple-and-list/array.mts#L37)
 - src/tuple-and-list/index-of-tuple.mts
     - [IndexOfTuple](./src/tuple-and-list/index-of-tuple.mts#L17)
     - [NegativeIndexOfTuple](./src/tuple-and-list/index-of-tuple.mts#L52)
+- src/tuple-and-list/length-constrained-tuple.mts
+    - [FixedLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L18)
+    - [MutableFixedLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L28)
+    - [MutableMinLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L47)
+    - [MinLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L62)
+    - [BoundedLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L109)
+    - [MutableBoundedLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L125)
+    - [MaxLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L149)
+    - [MutableMaxLengthTuple](./src/tuple-and-list/length-constrained-tuple.mts#L164)
 - src/tuple-and-list/list.mts
     - [List.Head](./src/tuple-and-list/list.mts#L19)
     - [List.Last](./src/tuple-and-list/list.mts#L32)
