+++ src/tuple-and-list/index.mts
@@ -1,5 +1,6 @@
 export * from './array.mjs';
 export * from './index-of-tuple.mjs';
+export * from './length-constrained-tuple.mjs';
 export * from './list.mjs';
 export * from './make-tuple.mjs';
 export * from './tuple.mjs';
